requirements.txt
streamlit
langchain
langchain-google-genai
pypdf2
python-dotenv
faiss-cpu
langchain-community
nest-asyncio